Move v1 client and structs to its own package.
This change prepares the introduction of structs and routines to
access Trusty's v2 endpoints. The idea is to eventually have users
import either `pkg/v1` or `pkg/v2` without ambiguity, while also
organizing the code a bit more intuitively.
