+++ pkg/parser/pomxml.go
@@ -18,7 +18,7 @@ package parser
 import (
 	"encoding/xml"
 
-	"github.com/stacklok/trusty-sdk-go/pkg/types"
+	"github.com/stacklok/trusty-sdk-go/pkg/v1/types"
 )
 
 // ParsePomXml parses the content of a POM XML file and returns a slice of dependencies
