+++ pkg/parser/requirements.go
@@ -18,7 +18,7 @@ package parser
 import (
 	"strings"
 
-	"github.com/stacklok/trusty-sdk-go/pkg/types"
+	"github.com/stacklok/trusty-sdk-go/pkg/v1/types"
 )
 
 // ParseRequirementsTxt parses requirements.txt content and extracts dependencies
